Refactored recoil mechanics for pistol and shotgun weapons by consolidating the applyRecoil method in the Weapon class. Removed redundant recoil logic from individual weapon classes and adjusted recoil application to use a strength scale parameter. This improves code maintainability and consistency across weapon types.
