+++ packages/game-server/src/entities/weapons/weapon.ts
@@ -7,6 +7,8 @@ import { WeaponKey } from "../../../../game-shared/src/util/inventory";
 import { Direction } from "../../../../game-shared/src/util/direction";
 import Vector2 from "@/util/vector2";
 import { WeaponConfig, weaponRegistry } from "@shared/entities";
+import type { IEntity } from "@/entities/types";
+import { applyWeaponRecoil } from "@/entities/util/recoil";
 
 export abstract class Weapon extends Entity {
   public static readonly Size = new Vector2(16, 16);
@@ -37,4 +39,20 @@ export abstract class Weapon extends Entity {
   ): void;
 
   public abstract getCooldown(): number;
+
+  protected applyRecoil(
+    player: IEntity,
+    facing: Direction,
+    aimAngle: number | undefined,
+    strengthScale: number = 1
+  ): void {
+    const recoilBase = this.getConfig().stats.recoilKnockback ?? 0;
+    const recoilStrength = recoilBase * strengthScale;
+    applyWeaponRecoil({
+      player,
+      facing,
+      aimAngle,
+      strength: recoilStrength,
+    });
+  }
 }
